+++ src/modules/order/refund-requests/producers/index.ts
@@ -1,0 +1,1 @@
+export * from './refund-request.producer';
