+++ src/queue/constants/queue-name.constant.ts
@@ -75,4 +75,4 @@ export const GIVE_REWARD_QUEUE = 'O2R-GIVE_REWARD';
 export const INDEX_ORDER_ITEMS_QUEUE = 'O2O-INDEX_ORDER_ITEMS';
 export const DELETE_ORDER_ITEMS_INDEX_QUEUE = 'O2O-DELETE_ORDER_ITEMS_INDEX';
 
-export const INDEX_REFUND_REQUEST_QUEUE = 'O2R-INDEX_REFUND_REQUEST';
+export const INDEX_REFUND_REQUEST_QUEUE = 'R2R-INDEX_REFUND_REQUEST';
